// src/generated/models/BuildTask.ts
export interface BuildTask {
    /**
     * 
     * @type {number}
     * @memberof BuildTask
     */
    readonly id: number;
    /**
     * 
     * @type {Date}
     * @memberof BuildTask
     */
    readonly createdAt: Date;
    /**
     * 
     * @type {BuildSettingCreatedBy}
     * @memberof BuildTask
     */
    createdBy: BuildSettingCreatedBy;
    /**
     * 
     * @type {Date}
     * @memberof BuildTask
     */
    readonly updatedAt: Date;
    /**
     * 
     * @type {BuildSettingCreatedBy}
     * @memberof BuildTask
     */
    updatedBy: BuildSettingCreatedBy;
    /**
     * 
     * @type {BuildTaskCommit}
     * @memberof BuildTask
     */
    commit: BuildTaskCommit;
    /**
     * 
     * @type {StateEnum}
     * @memberof BuildTask
     */
    readonly state: StateEnum | null;
    /**
     * 
     * @type {string}
     * @memberof BuildTask
     */
    version: string;
    /**
     * 
     * @type {string}
     * @memberof BuildTask
     */
    commitSha: string;
    /**
     * 
     * @type {string}
     * @memberof BuildTask
     */
    readonly yaml: string | null;
    /**
     * 
     * @type {string}
     * @memberof BuildTask
     */
    log?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BuildTask
     */
    uid?: string | null;
}